// app/RootComponents/WeatherChart/type.ts
export interface Root {
  city: string;
  date: string;
  value: Value;
}

export interface Value {
  temp: number;
  feels_like: number;
  temp_min: number;
  temp_max: number;
  pressure: number;
  sea_level: number;
  grnd_level: number;
  humidity: number;
  temp_kf: number;
  weather: Weather[];
  wind: Wind;
}

export interface Weather {
  id: number;
  main: string;
  description: string;
  icon: string;
}

export interface Wind {
  speed: number;
  deg: number;
  gust: number;
}